docs: for stops and routes data, improves function descriptions and explain how to use files
This commit only add descriptions of functions and how to use files
related to stops and routes data. Doesn't modify any functional code (in
this matter, it only removes call to three packages that were not used
in the fetch_all_data endpoint, not any other line of code is being
modified).